refactor(sub_info): rename `sub_dag.go` to `sub_info.go` & other changes based on pr review
